backend/requirements.txt
fastapi[standard]>=0.110.0
pydantic>=2.6.0
python-dotenv>=1.0.1
google-genai>=1.0.0
pdfplumber>=0.11.0
beautifulsoup4>=4.12.0
python-multipart>=0.0.9
